Move tests for partitioning modules to new files
Create new files for unit tests of the partitioning modules
and move the existing tests there.
