+++ tests/nosetests/pyanaconda_tests/module_storage_test.py
@@ -22,18 +22,12 @@ import tempfile
 import unittest
 from unittest.mock import patch, call, Mock
 
-from blivet.devicelibs.crypto import MIN_CREATE_ENTROPY
-from blivet.formats.luks import LUKS2PBKDFArgs
-
-from pykickstart.constants import AUTOPART_TYPE_LVM_THINP, AUTOPART_TYPE_PLAIN, AUTOPART_TYPE_LVM
-
 from pyanaconda.bootloader.grub2 import IPSeriesGRUB2, GRUB2
 from pyanaconda.bootloader.zipl import ZIPL
-from pyanaconda.dbus.typing import get_variant, Str, Bool, ObjPath
-from pyanaconda.modules.common.constants.objects import AUTO_PARTITIONING, MANUAL_PARTITIONING
+from pyanaconda.dbus.typing import ObjPath
+from pyanaconda.modules.common.constants.objects import AUTO_PARTITIONING
 from pyanaconda.modules.common.errors.configuration import StorageDiscoveryError
-from pyanaconda.modules.common.errors.storage import InvalidStorageError, UnavailableStorageError, \
-    UnavailableDataError
+from pyanaconda.modules.common.errors.storage import InvalidStorageError
 from pyanaconda.modules.common.task import TaskInterface
 from pyanaconda.modules.storage.dasd import DASDModule
 from pyanaconda.modules.storage.dasd.dasd_interface import DASDInterface
@@ -44,15 +38,6 @@ from pyanaconda.modules.storage.fcoe.discover import FCOEDiscoverTask
 from pyanaconda.modules.storage.fcoe.fcoe_interface import FCOEInterface
 from pyanaconda.modules.storage.installation import ActivateFilesystemsTask, MountFilesystemsTask, \
     WriteConfigurationTask
-from pyanaconda.modules.storage.partitioning import AutoPartitioningModule, \
-    ManualPartitioningModule, CustomPartitioningModule
-from pyanaconda.modules.storage.partitioning.automatic_interface import AutoPartitioningInterface
-from pyanaconda.modules.storage.partitioning.automatic_partitioning import \
-    AutomaticPartitioningTask
-from pyanaconda.modules.storage.partitioning.base_interface import PartitioningInterface
-from pyanaconda.modules.storage.partitioning.custom_partitioning import CustomPartitioningTask
-from pyanaconda.modules.storage.partitioning.manual_interface import ManualPartitioningInterface
-from pyanaconda.modules.storage.partitioning.manual_partitioning import ManualPartitioningTask
 from pyanaconda.modules.storage.partitioning.validate import StorageValidateTask
 from pyanaconda.modules.storage.reset import StorageResetTask
 from pyanaconda.modules.storage.storage import StorageModule
@@ -61,8 +46,7 @@ from pyanaconda.modules.storage.zfcp import ZFCPModule
 from pyanaconda.modules.storage.zfcp.discover import ZFCPDiscoverTask
 from pyanaconda.modules.storage.zfcp.zfcp_interface import ZFCPInterface
 from pyanaconda.storage.checker import StorageCheckerReport
-from pyanaconda.storage.initialization import create_storage
-from tests.nosetests.pyanaconda_tests import check_kickstart_interface, check_dbus_property
+from tests.nosetests.pyanaconda_tests import check_kickstart_interface
 
 
 class StorageInterfaceTestCase(unittest.TestCase):
@@ -877,269 +861,6 @@ class StorageTasksTestCase(unittest.TestCase):
         storage.reset.assert_called_once()
 
 
-class AutopartitioningInterfaceTestCase(unittest.TestCase):
-    """Test DBus interface of the auto partitioning module."""
-
-    def setUp(self):
-        """Set up the module."""
-        self.autopart_module = AutoPartitioningModule()
-        self.autopart_interface = AutoPartitioningInterface(self.autopart_module)
-
-    def _test_dbus_property(self, *args, **kwargs):
-        check_dbus_property(
-            self,
-            AUTO_PARTITIONING,
-            self.autopart_interface,
-            *args, **kwargs
-        )
-
-    def enabled_property_test(self):
-        """Test the property enabled."""
-        self._test_dbus_property(
-            "Enabled",
-            True
-        )
-
-    def type_property_test(self):
-        """Test the type property."""
-        self._test_dbus_property(
-            "Type",
-            AUTOPART_TYPE_LVM_THINP
-        )
-
-        self._test_dbus_property(
-            "Type",
-            AUTOPART_TYPE_PLAIN
-        )
-
-        self._test_dbus_property(
-            "Type",
-            AUTOPART_TYPE_LVM
-        )
-
-    def filesystem_type_property_test(self):
-        """Test the filesystem property."""
-        self._test_dbus_property(
-            "FilesystemType",
-            "ext4"
-        )
-
-    def nohome_property_test(self):
-        """Test the nohome property."""
-        def setter(value):
-            self.autopart_module.set_nohome(value)
-            self.autopart_module.module_properties_changed.emit()
-
-        self._test_dbus_property(
-            "NoHome",
-            True,
-            setter=setter
-        )
-
-    def noboot_property_test(self):
-        """Test the noboot property."""
-        def setter(value):
-            self.autopart_module.set_noboot(value)
-            self.autopart_module.module_properties_changed.emit()
-
-        self._test_dbus_property(
-            "NoBoot",
-            True,
-            setter=setter
-        )
-
-    def noswap_property_test(self):
-        """Test the noswap property."""
-        def setter(value):
-            self.autopart_module.set_noswap(value)
-            self.autopart_module.module_properties_changed.emit()
-
-        self._test_dbus_property(
-            "NoSwap",
-            True,
-            setter=setter
-        )
-
-    def encrypted_property_test(self):
-        """Test the encrypted property."""
-        self._test_dbus_property(
-            "Encrypted",
-            True
-        )
-
-    def cipher_property_test(self):
-        """Test the cipher property,"""
-        self._test_dbus_property(
-            "Cipher",
-            "aes-xts-plain64"
-        )
-
-    def passphrase_property_test(self):
-        """Test the passphrase property."""
-        self._test_dbus_property(
-            "Passphrase",
-            "123456"
-        )
-
-    def luks_version_property_test(self):
-        """Test the luks version property."""
-        self._test_dbus_property(
-            "LUKSVersion",
-            "luks1"
-        )
-
-    def pbkdf_property_test(self):
-        """Test the PBKDF property."""
-        self._test_dbus_property(
-            "PBKDF",
-            "argon2i"
-        )
-
-    def pbkdf_memory_property_test(self):
-        """Test the PBKDF memory property."""
-        self._test_dbus_property(
-            "PBKDFMemory",
-            256
-        )
-
-    def pbkdf_time_property_test(self):
-        """Test the PBKDF time property."""
-        self._test_dbus_property(
-            "PBKDFTime",
-            100
-        )
-
-    def pbkdf_iterations_property_test(self):
-        """Test the PBKDF iterations property."""
-        self._test_dbus_property(
-            "PBKDFIterations",
-            1000
-        )
-
-    def escrowcert_property_test(self):
-        """Test the escrowcert property."""
-        self._test_dbus_property(
-            "Escrowcert",
-            "file:///tmp/escrow.crt"
-        )
-
-    def backup_passphrase_enabled_property_test(self):
-        """Test the backup passphrase enabled property."""
-        self._test_dbus_property(
-            "BackupPassphraseEnabled",
-            True
-        )
-
-    def pbkdf_args_test(self):
-        """Test the pbkdf_args property."""
-        self.autopart_module.set_encrypted(False)
-        self.assertEqual(self.autopart_module.pbkdf_args, None)
-
-        self.autopart_module.set_encrypted(True)
-        self.autopart_module.set_luks_version("luks1")
-        self.assertEqual(self.autopart_module.pbkdf_args, None)
-
-        self.autopart_module.set_encrypted(True)
-        self.autopart_module.set_luks_version("luks2")
-        self.assertEqual(self.autopart_module.pbkdf_args, None)
-
-        self.autopart_module.set_encrypted(True)
-        self.autopart_module.set_luks_version("luks2")
-        self.autopart_module.set_pbkdf("argon2i")
-        self.autopart_module.set_pbkdf_memory(256)
-        self.autopart_module.set_pbkdf_iterations(1000)
-        self.autopart_module.set_pbkdf_time(100)
-
-        pbkdf_args = self.autopart_module.pbkdf_args
-        self.assertIsInstance(pbkdf_args, LUKS2PBKDFArgs)
-        self.assertEqual(pbkdf_args.type, "argon2i")
-        self.assertEqual(pbkdf_args.max_memory_kb, 256)
-        self.assertEqual(pbkdf_args.iterations, 1000)
-        self.assertEqual(pbkdf_args.time_ms, 100)
-
-    def luks_format_args_test(self):
-        """Test the luks_format_args property."""
-        storage = create_storage()
-        storage.encryption_passphrase = "default"
-        storage._escrow_certificates["file:///tmp/escrow.crt"] = "CERTIFICATE"
-        self.autopart_module.on_storage_reset(storage)
-
-        self.autopart_module.set_encrypted(False)
-        self.assertEqual(self.autopart_module.luks_format_args, {})
-
-        self.autopart_module.set_encrypted(True)
-        self.assertEqual(self.autopart_module.luks_format_args, {
-            "passphrase": "default",
-            "cipher": "",
-            "luks_version": "luks2",
-            "pbkdf_args": None,
-            "escrow_cert": None,
-            "add_backup_passphrase": False,
-            "min_luks_entropy": MIN_CREATE_ENTROPY,
-        })
-
-        self.autopart_module.set_encrypted(True)
-        self.autopart_module.set_luks_version("luks1")
-        self.autopart_module.set_passphrase("passphrase")
-        self.autopart_module.set_cipher("aes-xts-plain64")
-        self.autopart_module.set_escrowcert("file:///tmp/escrow.crt")
-        self.autopart_module.set_backup_passphrase_enabled(True)
-        self.assertEqual(self.autopart_module.luks_format_args, {
-            "passphrase": "passphrase",
-            "cipher": "aes-xts-plain64",
-            "luks_version": "luks1",
-            "pbkdf_args": None,
-            "escrow_cert": "CERTIFICATE",
-            "add_backup_passphrase": True,
-            "min_luks_entropy": MIN_CREATE_ENTROPY,
-        })
-
-    def reset_test(self):
-        """Test the reset of the storage."""
-        with self.assertRaises(UnavailableStorageError):
-            if self.autopart_module.storage:
-                self.fail("The storage shouldn't be available.")
-
-        storage = Mock()
-        self.autopart_module.on_storage_reset(storage)
-
-        self.assertEqual(self.autopart_module._current_storage, storage)
-        self.assertIsNone(self.autopart_module._storage_playground)
-
-        self.assertNotEqual(self.autopart_module.storage, storage)
-        self.assertIsNotNone(self.autopart_module._storage_playground)
-
-    @patch('pyanaconda.dbus.DBus.publish_object')
-    def configure_with_task_test(self, publisher):
-        """Test ConfigureWithTask."""
-        self.autopart_module.on_storage_reset(Mock())
-        task_path = self.autopart_interface.ConfigureWithTask()
-
-        publisher.assert_called_once()
-        object_path, obj = publisher.call_args[0]
-
-        self.assertEqual(task_path, object_path)
-        self.assertIsInstance(obj, TaskInterface)
-
-        self.assertIsInstance(obj.implementation, AutomaticPartitioningTask)
-        self.assertEqual(obj.implementation._storage, self.autopart_module.storage)
-
-    @patch('pyanaconda.dbus.DBus.publish_object')
-    def validate_with_task_test(self, publisher):
-        """Test ValidateWithTask."""
-        self.autopart_module.on_storage_reset(Mock())
-        task_path = self.autopart_interface.ValidateWithTask()
-
-        publisher.assert_called_once()
-        object_path, obj = publisher.call_args[0]
-
-        self.assertEqual(task_path, object_path)
-        self.assertIsInstance(obj, TaskInterface)
-
-        self.assertIsInstance(obj.implementation, StorageValidateTask)
-        self.assertEqual(obj.implementation._storage, self.autopart_module.storage)
-
-
 class DASDInterfaceTestCase(unittest.TestCase):
     """Test DBus interface of the DASD module."""
 
@@ -1351,211 +1072,6 @@ class ZFCPTasksTestCase(unittest.TestCase):
         sanitized_lun = blockdev.s390.zfcp_sanitize_lun_input.return_value
 
         zfcp.add_fcp.assert_called_once_with(sanitized_dev, sanitized_wwpn, sanitized_lun)
-
-
-class ManualPartitioningInterfaceTestCase(unittest.TestCase):
-    """Test DBus interface of the manual partitioning module."""
-
-    def setUp(self):
-        """Set up the module."""
-        self.manual_part_module = ManualPartitioningModule()
-        self.manual_part_interface = ManualPartitioningInterface(self.manual_part_module)
-
-    def _test_dbus_property(self, *args, **kwargs):
-        check_dbus_property(
-            self,
-            MANUAL_PARTITIONING,
-            self.manual_part_interface,
-            *args, **kwargs
-        )
-
-    def enabled_property_test(self):
-        """Test the enabled property."""
-        self._test_dbus_property(
-            "Enabled",
-            True
-        )
-
-        self._test_dbus_property(
-            "Enabled",
-            False
-        )
-
-    def mount_points_property_test(self):
-        """Test the mount points property."""
-        self._test_dbus_property(
-            "Requests",
-            []
-        )
-
-        in_value = [
-            {
-                "mount-point": "/boot",
-                "device-spec": "/dev/sda1"
-            }
-        ]
-
-        out_value = [
-            {
-                "mount-point": get_variant(Str, "/boot"),
-                "device-spec": get_variant(Str, "/dev/sda1"),
-                "reformat": get_variant(Bool, False),
-                "format-type": get_variant(Str, ""),
-                "format-options": get_variant(Str, ""),
-                "mount-options": get_variant(Str, "")
-            }
-        ]
-
-        self._test_dbus_property(
-            "Requests",
-            in_value,
-            out_value
-        )
-
-        in_value = [
-            {
-                "mount-point":  "/boot",
-                "device-spec": "/dev/sda1",
-                "reformat": True,
-                "format-type": "xfs",
-                "format-options": "-L BOOT",
-                "mount-options": "user"
-            }
-        ]
-
-        out_value = [
-            {
-                "mount-point": get_variant(Str, "/boot"),
-                "device-spec": get_variant(Str, "/dev/sda1"),
-                "reformat": get_variant(Bool, True),
-                "format-type": get_variant(Str, "xfs"),
-                "format-options": get_variant(Str, "-L BOOT"),
-                "mount-options": get_variant(Str, "user")
-            }
-        ]
-
-        self._test_dbus_property(
-            "Requests",
-            in_value,
-            out_value,
-        )
-
-        in_value = [
-            {
-                "mount-point": "/boot",
-                "device-spec": "/dev/sda1"
-            },
-            {
-                "mount-point": "/",
-                "device-spec": "/dev/sda2",
-                "reformat": True
-            }
-        ]
-
-        out_value = [
-            {
-                "mount-point": get_variant(Str, "/boot"),
-                "device-spec": get_variant(Str, "/dev/sda1"),
-                "reformat": get_variant(Bool, False),
-                "format-type": get_variant(Str, ""),
-                "format-options": get_variant(Str, ""),
-                "mount-options": get_variant(Str, "")
-            },
-            {
-                "mount-point": get_variant(Str, "/"),
-                "device-spec": get_variant(Str, "/dev/sda2"),
-                "reformat": get_variant(Bool, True),
-                "format-type": get_variant(Str, ""),
-                "format-options": get_variant(Str, ""),
-                "mount-options": get_variant(Str, "")
-            }
-        ]
-
-        self._test_dbus_property(
-            "Requests",
-            in_value,
-            out_value
-        )
-
-    @patch('pyanaconda.dbus.DBus.publish_object')
-    def configure_with_task_test(self, publisher):
-        """Test ConfigureWithTask."""
-        self.manual_part_module.on_storage_reset(Mock())
-        task_path = self.manual_part_interface.ConfigureWithTask()
-
-        publisher.assert_called_once()
-        object_path, obj = publisher.call_args[0]
-
-        self.assertEqual(task_path, object_path)
-        self.assertIsInstance(obj, TaskInterface)
-
-        self.assertIsInstance(obj.implementation, ManualPartitioningTask)
-        self.assertEqual(obj.implementation._storage, self.manual_part_module.storage)
-
-    @patch('pyanaconda.dbus.DBus.publish_object')
-    def validate_with_task_test(self, publisher):
-        """Test ValidateWithTask."""
-        self.manual_part_module.on_storage_reset(Mock())
-        task_path = self.manual_part_interface.ValidateWithTask()
-
-        publisher.assert_called_once()
-        object_path, obj = publisher.call_args[0]
-
-        self.assertEqual(task_path, object_path)
-        self.assertIsInstance(obj, TaskInterface)
-
-        self.assertIsInstance(obj.implementation, StorageValidateTask)
-        self.assertEqual(obj.implementation._storage, self.manual_part_module.storage)
-
-
-class CustomPartitioningInterfaceTestCase(unittest.TestCase):
-    """Test DBus interface of the custom partitioning module."""
-
-    def setUp(self):
-        """Set up the module."""
-        self.custom_part_module = CustomPartitioningModule()
-        self.custom_part_interface = PartitioningInterface(self.custom_part_module)
-
-    def data_test(self, ):
-        """Test the data property."""
-        with self.assertRaises(UnavailableDataError):
-            if self.custom_part_module.data:
-                self.fail("The data should not be available.")
-
-        data = Mock()
-        self.custom_part_module.process_kickstart(data)
-        self.assertEqual(self.custom_part_module.data, data)
-
-    @patch('pyanaconda.dbus.DBus.publish_object')
-    def configure_with_task_test(self, publisher):
-        """Test ConfigureWithTask."""
-        self.custom_part_module.on_storage_reset(Mock())
-        self.custom_part_module.process_kickstart(Mock())
-        task_path = self.custom_part_interface.ConfigureWithTask()
-
-        publisher.assert_called_once()
-        object_path, obj = publisher.call_args[0]
-
-        self.assertEqual(task_path, object_path)
-        self.assertIsInstance(obj, TaskInterface)
-
-        self.assertIsInstance(obj.implementation, CustomPartitioningTask)
-        self.assertEqual(obj.implementation._storage, self.custom_part_module.storage)
-
-    @patch('pyanaconda.dbus.DBus.publish_object')
-    def validate_with_task_test(self, publisher):
-        """Test ValidateWithTask."""
-        self.custom_part_module.on_storage_reset(Mock())
-        task_path = self.custom_part_interface.ValidateWithTask()
-
-        publisher.assert_called_once()
-        object_path, obj = publisher.call_args[0]
-
-        self.assertEqual(task_path, object_path)
-        self.assertIsInstance(obj, TaskInterface)
-
-        self.assertIsInstance(obj.implementation, StorageValidateTask)
-        self.assertEqual(obj.implementation._storage, self.custom_part_module.storage)
 
 
 class StorageValidationTasksTestCase(unittest.TestCase):
